Create API Key Partner Exception Report
Use 1x per month to see if Sub API keys are not properly linked to a partner account, but the main keys are. This is due to partner flagging able to be done at the account or opportunity level. 

Issue is that after the 8th of the month, historical finance numbers are locked, which could cause revenue to not be associated to a partner.